Split "Generate function" intention by two sub-intentions
To improve UX
