+++ src/main/resources/intentionDescriptions/ExtractExpressionToFunctionIntention/description.html
@@ -1,0 +1,5 @@
+<html lang="en">
+<body>
+Extracts expression to a function.
+</body>
+</html>
