+++ src/main/resources/intentionDescriptions/GenerateFunctionFromGoalIntention/after.ard.template
