+++ src/main/resources/intentionDescriptions/GenerateFunctionFromGoalIntention/before.ard.template
