+++ src/main/resources/intentionDescriptions/GenerateFunctionFromGoalIntention/description.html
