+++ src/main/resources/messages/ArendBundle.properties
@@ -41,3 +41,5 @@ arend.inspection.redundant.parentheses.message=Redundant parentheses
 arend.unwrap.parentheses.fix=Unwrap parentheses
 
 arend.generate.function=Generate function
+arend.generate.function.from.goal=Generate function from goal
+arend.generate.function.from.expression=Extract expression to function
